requirements/generator.txt
nbformat
pyyaml
pymongo
jinja2 